Remove silent handler from cluster object
PR #489 fixed the issue where any Golang function request to the
cluster object would be handled by an in-memory implementation instead
of having the system panic due to having no implementation available.

This could introduce another issue in that it could have a 'silent'
handler which could be unexpected. Instead, this fix makes it so
that the cluster object has a handler which always returns error
on default, removing the silent handler.

Signed-off-by: Luis Pabón <lpabon@gmail.com>