HTTPS and Auth support for the SDK
Previously the REST Gateway would directly connect to the gRPC
port. Now with HTTPS/TLS support, this is more complicated since
the CERTS would have a CN not labeled 'localhost'.

For simplicity the SDK now also has a unix domain socket. This
UDS is used for local access and for the REST Gateway to communicate
with the gRPC server.

All models support token authentication and authorization based
on JWTs.

All SDK unit tests run with HTTPS enabled.

Signed-off-by: Luis Pabón <luis@portworx.com>