Docker Volume API Cookie support
Docker calls some API functions like get() and path() before
calling the actual function to do the work. This makes the token
available to the first call, but not to the work function.

This fix provides a cookie for the volume, where it saves
the token for a time, so that it can be used for the actual function
that does the work.

Signed-off-by: Luis Pabón <luis@portworx.com>