SDK InspectWithFilters and Deep.Inspect
Inspect may be a high latency call in drivers, therefore doing
an sdk Enumerate/Inspect combination may be problematic. Instead,
this patch introduces a "deep" option for SDK Inspect() which makes it
act like it used to, but now instead of calling driver.Inspect(),
it calls the driver's Enumerate() calls, which may be cache friendly.

Signed-off-by: Luis Pabón <luis@portworx.com>